Добавляем миграцию с полнотекстным поиском
