--- Infrastructure/Store.Data.EF/ServiceCollectionExtensions.cs
@@ -1,26 +1,0 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Microsoft.EntityFrameworkCore;
-
-namespace Store.Data.EF
-{
-    public static class ServiceCollectionExtensions
-    {
-        public static IServiceCollection AddEfRepositories(this IServiceCollection services, string connectionString) 
-        {
-            services.AddDbContext<StoreDbContext>
-                (
-                    options => { options.UseSqlServer(connectionString); },
-                    ServiceLifetime.Transient
-                );
-            services.AddSingleton<IBookRepository, BookRepository>();
-            services.AddSingleton<IOrderRepository, OrderRepository>();
-
-            return services;
-        }
-    }
-}
