+++ Infrastructure/Store.Data.EF/StoreDbConteхt.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Internal;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
 using Newtonsoft.Json;
 using System;
@@ -372,6 +373,22 @@ namespace Store.Data.EF
             });
         }
     }
-    
+
+    public static class ServiceCollectionExtensions
+    {
+        public static IServiceCollection AddEfRepositories(this IServiceCollection services, string connectionString)
+        {
+            services.AddDbContext<StoreDbContext>
+                (
+                    options => { options.UseSqlServer(connectionString); },
+                    ServiceLifetime.Transient
+                );
+            services.AddSingleton<IBookRepository, BookRepository>();
+            services.AddSingleton<IOrderRepository, OrderRepository>();
+
+            return services;
+        }
+    }
+
 }
